Refactor: Break subscription UI into multiple files
- I moved helper functions from `subscriptions-ui.js` to `subscription-helpers.js`.
- I moved modal rendering, event handling, and data fetching logic from `subscriptions-ui.js` to `subscription-modal.js`.
- I moved global variables related to modal state into `subscription-modal.js`.
- I moved the generic `fetchAPI` function from `subscriptions-ui.js` to `utils.js`.
- I updated `subscriptions-ui.js` to be a minimal initializer for the subscription modal.
- I updated `main.js` to import and initialize the subscription UI through `subscriptions-ui.js`.
- I ensured all imports, exports, and function calls are updated correctly across the new file structure.
